+++ cmd/server.go
@@ -6,7 +6,14 @@ import (
 	"github.com/spf13/cobra"
 
 	"fmt"
+	"net"
 	"net/http"
+)
+
+const (
+	// Local server configuration
+	serverGraphQLPath = "/graphql"
+	serverLocalPort   = 8080
 )
 
 var serverCmd = &cobra.Command{
@@ -18,12 +25,47 @@ var serverCmd = &cobra.Command{
 		if err != nil {
 			return err
 		}
-		return startAPIServer(cfg)
+		return startAPIServer(cfg, nil)
 	},
 }
 
-func startAPIServer(cfg *config.AConfig) error {
-	fmt.Println("API server listening at :8080 ... ")
-	http.Handle("/graphql", server.GraphQLHandler(cfg))
-	return http.ListenAndServe(":8080", nil)
+// Function is called before booting the server to configure
+// additional routes for mux, and to provide "ready" hooks
+type configFunc func(*http.ServeMux, chan bool) error
+
+func startAPIServer(cfg *config.AConfig, configFn configFunc) error {
+	// Simple router, for now
+	mux := http.NewServeMux()
+
+	// GraphQL API is handled via API
+	mux.Handle(serverGraphQLPath, server.GraphQLHandler(cfg))
+
+	// Append additional routes (e.g. preview)
+	var ready chan bool = nil
+	if configFn != nil {
+		ready = make(chan bool)
+		if err := configFn(mux, ready); err != nil {
+			return err
+		}
+	}
+
+	// Initialize server
+	s := &http.Server{Handler: mux}
+	s.Addr = fmt.Sprintf(":%d", serverLocalPort)
+
+	// Open port for listening
+	l, err := net.Listen("tcp", s.Addr)
+	if err != nil {
+		return err
+	}
+
+	// Notify listeners if ready (see preview command)
+	if ready != nil {
+		ready <- true
+		close(ready)
+	}
+
+	// Serve server API
+	fmt.Printf("API server listening at %s ... \n", s.Addr)
+	return s.Serve(l)
 }