+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Switched to Goldmark for Markdown rendering
 - Campaign subject is now also a template
 - Removed usage of global configurations
+- Preview UI proxying & config
 
 ## 0.2.0 (25 Jul 2017)